feat: add user authentication and order management endpoints with Swagger documentation
- Updated package.json to include swagger-jsdoc and swagger-ui-express for API documentation.
- Created initial database migration for Users, Order, and Items tables.
- Modified Prisma schema to define Users, Order, and Items models with relationships.
- Refactored auth.routes.js to implement user registration and login with Swagger annotations.
- Enhanced order.routes.js to manage orders with Swagger documentation and user authentication.
- Updated auth.schema.js to define validation schemas for user registration and login.
- Refactored auth.service.js to throw AppError for authentication errors.
- Enhanced order.service.js to include user validation for order operations and format output.
- Added Swagger configuration in swagger.js for API documentation.
- Created AppError utility for consistent error handling across services.

+++ service/order.service.js
@@ -1,48 +1,86 @@
 import prisma from '../config/database.js';
+import { AppError } from '../utils/AppError.js';
 
-export const createOrder = async (orderDTO) => {
+
+const formatOrderOutput = (order) => {
+    delete order.userId;
+    order.items = order.items.map(item => {
+        const { orderId, ...rest } = item;
+        return rest;
+    });
+    return order;
+}
+
+export const createOrder = async (orderDTO, req) => {
+
     const orderData = {
         orderId: orderDTO.numeroPedido,
+        userId: req.user.id,
         value: orderDTO.valorTotal,
         creationDate: new Date(orderDTO.dataCriacao),
         items: {
             create: orderDTO.items.map(item => ({
-                productId: parseInt(item.idItem),
+                productId: item.idItem,
                 quantity: item.quantidadeItem,
                 price: item.valorItem
             }))
         }
     };
 
-    return await prisma.order.create({
+    const order = await prisma.order.create({
         data: orderData,
         include: { items: true }
     });
+
+    return formatOrderOutput(order);
 };
 
-export const getOrderById = async (orderId) => {
+export const getOrderById = async (orderId, req) => {
     const order = await prisma.order.findUnique({
         where: { orderId },
         include: { items: true }
     });
 
     if (!order) {
-        const error = new Error('Pedido não encontrado');
-        error.statusCode = 404;
-        throw error;
+        throw new AppError("Pedido não encontrado", 404)
     }
 
-    return order;
+    if (order.userId !== req.user.id) {
+        throw new AppError("Ação não permitida", 401)
+    }
+
+    return formatOrderOutput(order);
 };
 
-export const getAllOrders = async () => {
-    return await prisma.order.findMany({
-        include: { items: true },
-        orderBy: { creationDate: 'desc' }
+export const getAllOrders = async (req) => {
+    const orders = await prisma.order.findMany({
+        where: {
+            userId: req.user.id
+        },
+        include: {
+            items: true
+        },
+        orderBy: {
+            creationDate: 'desc'
+        }
     });
+
+    return orders.map((order) => formatOrderOutput(order))
 };
 
-export const updateOrder = async (orderId, updateDTO) => {
+export const updateOrder = async (orderId, updateDTO, req) => {
+    const order = await prisma.order.findUnique({
+        where: { orderId }
+    });
+
+    if (!order) {
+        throw new AppError("Pedido não encontrado", 404)
+    }
+
+    if (order.userId !== req.user.id) {
+        throw new AppError("Você não tem permissão para alterar esse pedido", 403)
+    }
+
     const updateData = {};
 
     if (updateDTO.valorTotal !== undefined) {
@@ -57,22 +95,40 @@ export const updateOrder = async (orderId, updateDTO) => {
         updateData.items = {
             deleteMany: {},
             create: updateDTO.items.map(item => ({
-                productId: parseInt(item.idItem),
+                productId: item.idItem,
                 quantity: item.quantidadeItem,
                 price: item.valorItem
             }))
         };
     }
 
-    return await prisma.order.update({
+    const updatedOrder = await prisma.order.update({
         where: { orderId },
         data: updateData,
         include: { items: true }
     });
+
+    return formatOrderOutput(updatedOrder);
 };
 
-export const deleteOrder = async (orderId) => {
-    return await prisma.order.delete({
+export const deleteOrder = async (orderId, req) => {
+
+    const order = await prisma.order.findUnique({
+        where: { orderId },
+        select: { userId: true }
+    });
+
+    if (!order) {
+        throw new AppError("Pedido não encontrado", 404)
+    }
+
+    if (order.userId !== req.user.id) {
+        throw new AppError("Você não tem permissão para deletar esse pedido", 403)
+    }
+
+    const deleted = await prisma.order.delete({
         where: { orderId }
     });
+
+    return deleted;
 };
